+++ src/main/java/cart/exception/ControllerExceptionHandler.java
@@ -1,10 +1,6 @@
-package cart.ui;
+package cart.exception;
 
-import cart.dto.response.ExceptionResponse;
-import cart.exception.AuthenticationException;
-import cart.exception.CartItemException;
-import cart.exception.MemberException;
-import cart.exception.ProductException;
+import java.util.stream.Collectors;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -14,8 +10,6 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import java.util.stream.Collectors;
 
 @ControllerAdvice
 public class ControllerExceptionHandler extends ResponseEntityExceptionHandler {
