+++ src/main/java/cart/ui/api/CartItemApiController.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.ui.api;
 
 import cart.application.CartItemService;
 import cart.configuration.resolver.Checkout;
