+++ src/main/java/cart/ui/api/ProductApiController.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.ui.api;
 
 import cart.application.ProductService;
 import cart.dto.request.ProductRequest;
