+++ src/test/java/cart/ui/api/DocsTest.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.ui.api;
 
 import cart.application.CartItemService;
 import cart.application.OrderService;
