+++ src/test/java/cart/ui/api/OrderApiControllerTest.java
@@ -1,8 +1,8 @@
-package cart.ui;
+package cart.ui.api;
 
 import static cart.fixture.DomainFixture.CHICKEN;
 import static cart.fixture.DomainFixture.MEMBER_A;
-import static cart.ui.RestDocsConfiguration.field;
+import static cart.ui.api.RestDocsConfiguration.field;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyList;
@@ -32,6 +32,7 @@ import cart.domain.OrderItem;
 import cart.dto.request.OrderRequest;
 import cart.dto.response.OrderResponse;
 import cart.dto.response.SortedOrdersResponse;
+import cart.exception.ControllerExceptionHandler;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
