+++ src/test/java/cart/ui/api/RestDocsConfiguration.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.ui.api;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
